refactor(sources): add shared schema.go for schema.org JobPosting decode
Home the schema.org JobPosting helpers in one file: move schemaEmploymentType
out of northstone.go (and its test into schema_test.go), and add schemaAddress/
schemaPlace plus schemaPlaces (the single-or-array jobLocation decoder, promoted
from icims.go's icimsPlaces). icims/jobvite/twogis now decode through schemaPlaces.
Location-building stays per-adapter (icims filters UNAVAILABLE, jobvite/twogis
join locality+country only), so no behaviour change.
