+++ internal/sources/northstone_test.go
@@ -214,16 +214,3 @@ func TestNorthstoneRegisteredInAll(t *testing.T) {
 		t.Error("FilterableProviders() should include northstone (board-based)")
 	}
 }
-
-func TestSchemaEmploymentType(t *testing.T) {
-	cases := map[string]string{
-		"FULL_TIME": "full_time", "PART_TIME": "part_time",
-		"CONTRACTOR": "contract", "TEMPORARY": "contract",
-		"INTERN": "internship", "OTHER": "", "": "",
-	}
-	for in, want := range cases {
-		if got := schemaEmploymentType(in); got != want {
-			t.Errorf("schemaEmploymentType(%q) = %q, want %q", in, got, want)
-		}
-	}
-}
